Added Factory Class + Changed Difficulty Enums (#32)
* Added Chooser Factory Class

We now have a skeleton of a class that returns the user's choice of
difficulty.

* Changed Difficulty Enumeration

Now we have Random, Smart and Genius. I also changed all instances of
these enum values in the code.

* Integrated Factory to Computer Player

Now the computer player will change related to user choice

* Iss 31 (#34)

* Start ML implementation

* Use pair for sorting

* Fixed sorting function

* Prepare for histogram implementation

* Updated histogram psuedocode

* Added DecideMove pseudologic

* Added directory creation functionality

* Added frequency file readers

* Finished untested implementation

* Added initial write freq

* Removed pseudocode

* Added strerr compatibility

* Added Chooser Factory Class

We now have a skeleton of a class that returns the user's choice of
difficulty.

* Changed Difficulty Enumeration

Now we have Random, Smart and Genius. I also changed all instances of
these enum values in the code.

* Changed enum names