Added Multipanel GUI for EveryReceipt (#52)
* Start boilerplate

* Added Game Buttons with Start Buttons

Forgot that I was on the wrong branch. Thanks don

* Added Settings Option String Parsing

Now we can parse the enum value to a string. This is useful for buttons
to display a difficulty value.

* Started messing with visibility

* Moved all game text to one panel

Now we are having to change spacing for grid

* Fixed Segmentation Error

Accidentally removed the variable button_chosen_text. This caused seg
faults when we called a function that handled a move choice.

* Gonna change modifying sizers to modifying panels

* here we go...

* An Attempt of adding instance variable panels

Now the panels should have their own Generate function.

* It worked. Moving onto enable logic.

* Added Way to keep track of difficulty

the instance variable current_difficulty defaults to random, but later
changes to whatever the user picks in the start menu.

* Haha i am evan

I deleted a game reference comments

* comment cleanup